feat: Bill model edited; BillSerializer, BillView and urls of bill added

+++ API/models/Bills.py
@@ -8,6 +8,6 @@ from base_model.models import AbstractModel
 class Bills(AbstractModel):
     client = models.ForeignKey(verbose_name=_('Client'), to=Clients, on_delete=models.CASCADE)
     company_name = models.CharField(verbose_name=_('Company name'), max_length=50)
-    nit = models.PositiveIntegerField(verbose_name=_('NIT'))
+    nit = models.PositiveBigIntegerField(verbose_name=_('NIT'))
     code = models.PositiveSmallIntegerField(verbose_name=_('Verification code'), null=True)
     products = models.ManyToManyField(verbose_name=_('Products'), to=Products)
